Organize Get-Package Apex tests into dedicated test case
